+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update Dependencies [#1291](https://github.com/open-apparel-registry/open-apparel-registry/pull/1303)
 - Update facility details API [#1333](https://github.com/open-apparel-registry/open-apparel-registry/pull/1333)
 - Show footer in embed mode [#1339](https://github.com/open-apparel-registry/open-apparel-registry/pull/1339)
+- Remove "show other contributor info" embed option and hide other contributor information in embed mode [#1342](https://github.com/open-apparel-registry/open-apparel-registry/pull/1342)
 
 ### Deprecated